fix: レビュー指摘事項の対応
- CSP: 本番環境では unsafe-eval を除去（NODE_ENV で分岐）
- memberRemove: 全メンバーシップ削除時に primaryAttribute を participant にリセット
- OAuth state: timingSafeEqual による比較に変更
- verifyQrToken を session.ts から qrToken.ts に移動（責務の分離）
- /api/scan: proxy.ts と route handler の二重認証の設計意図をコメントで明記
- clientIp: プロキシトポロジの前提をコメントで明記
- dmSendLogs take:10 の根拠コメント追加
